Simplify and consolidate usage instructions
Before we described two different ways of running SourceCred (with yarn and running sourcecred directly). For
the purposes of the example instance, users should pretty much only be using the yarn commands. Also moved the instructions up
a bit so they dont get lost in the huge wall of text for plugin config.