Upgrade to beta-12
This upgrades the version of SourceCred being used. This also requires
an update to reflect that `sourcecred admin` was renamed to `sourcecred
serve`.

Test plan: Run the following commands:

```
yarn clean && \
yarn load && \
yarn graph && \
yarn score && \
yarn site && \
yarn serve
```

And verify that the site served is functional.

Note: As of beta-11, we now automatically add identities to the ledger
when we run `sourcecred graph`. For now I am not committing the updated
data/ledger.json file, we'll want to decide how to handle it (do we
commit a messy ledger.json file, and tell users to reset it when they
set up their instance?)