Add SourceCred as a default dependency (#9)
This commit adds SourceCred as a default dependency with a weight of
0.05.

Test plan: Regenerate the instance and load the frontend; you'll see
that there's a SourceCred identity with about 5% of the total Cred.

Note: Running the test plan will modify the `config/dependencies.json`,
because it will choose an id for SourceCred. We need to think about how
to ensure the users get a clean state for their own instance, c.f.
discussion in #8.

This also upgrades to sc beta-14, which has a necessary bugfix for the
dependency minted cred computation.